+++ crates/cli/tests/reference/targets-target-esm-mvp.bg.js
@@ -1,0 +1,21 @@
+let wasm;
+let wasmModule;
+export function __wbg_set_wasm(exports, module) {
+    wasm = exports;
+    wasmModule = module;
+}
+
+/**
+ * @param {number} a
+ * @param {number} b
+ * @returns {number}
+ */
+export function add_that_might_fail(a, b) {
+    const ret = wasm.add_that_might_fail(a, b);
+    return ret >>> 0;
+}
+
+export function __wbg_random_9526caf33df4270d() {
+    const ret = Math.random();
+    return ret;
+};