+++ crates/cli/tests/reference/targets-target-web-atomics.js
@@ -72,15 +72,15 @@ async function __wbg_load(module, imports) {
 
 function __wbg_get_imports(memory) {
     const imports = {};
-    imports.wbg = {};
-    imports.wbg.__wbg___wbindgen_throw_dd24417ed36fc46e = function(arg0, arg1) {
+    imports["./reference_test_bg.js"] = {};
+    imports["./reference_test_bg.js"].__wbg___wbindgen_throw_dd24417ed36fc46e = function(arg0, arg1) {
         throw new Error(getStringFromWasm0(arg0, arg1));
     };
-    imports.wbg.__wbg_random_e2b253f0e987bd7c = function() {
+    imports["./reference_test_bg.js"].__wbg_random_e2b253f0e987bd7c = function() {
         const ret = Math.random();
         return ret;
     };
-    imports.wbg.__wbindgen_init_externref_table = function() {
+    imports["./reference_test_bg.js"].__wbindgen_init_externref_table = function() {
         const table = wasm.__wbindgen_externrefs;
         const offset = table.grow(4);
         table.set(0, undefined);
@@ -89,7 +89,7 @@ function __wbg_get_imports(memory) {
         table.set(offset + 2, true);
         table.set(offset + 3, false);
     };
-    imports.wbg.memory = memory || new WebAssembly.Memory({initial:18,maximum:16384,shared:true});
+    imports["./reference_test_bg.js"].memory = memory || new WebAssembly.Memory({initial:18,maximum:16384,shared:true});
 
     return imports;
 }